feat: refactor breadcrumbs and add to them margin, and delete bad margintop in pages
