+++ src/modules/shared/components/Breadcrumbs/Breadcrumbs.tsx
@@ -1,47 +1,23 @@
-import styles from './Breadcrumbs.module.scss';
 import { Link, useLocation } from 'react-router-dom';
+import styles from './Breadcrumbs.module.scss';
+import { Crumb } from './Crumb/Crumb';
 
 export const Breadcrumbs = () => {
   const location = useLocation();
-  let currentLink = '';
 
-  const upFirstLetterOfTheWord = (word: string) => {
-    return word.charAt(0).toUpperCase() + word.slice(1);
-  };
-
-  const crumbs = location.pathname
-    .split('/')
-    .filter(crumb => crumb !== '')
-    .map((crumb, index, array) => {
-      currentLink += `/${crumb}`;
-
-      return (
-        <div
-          className={
-            index === array.length - 1
-              ? styles['breadcrumbs__crumb-active']
-              : styles['breadcrumbs__crumb']
-          }
-          key={crumb}
-        >
-          <Link to={currentLink}>{upFirstLetterOfTheWord(crumb)}</Link>
-          {index < array.length - 1 && (
-            <span className={styles['separator']}></span>
-          )}
-        </div>
-      );
-    });
+  const crumbs = location.pathname.split('/').filter(crumb => crumb !== '');
 
   return (
     <div className={styles['breadcrumbs']}>
-      <Link className={styles['breadcrumbs__crumb']} to="/">
+      <Link to="/">
         <img
-          className={styles.icon}
+          className={styles['breadcrumbs__icon-home']}
           src="images/original/breadcrumbs/home.svg"
+          alt="Home"
         />
       </Link>
       <span className={styles['separator']}></span>
-      {crumbs}
+      <Crumb crumbs={crumbs} />
     </div>
   );
 };
